Include list of tests via RequireJS
WHen running many tests you should be able to include them as a separate file
as per issue #1. This will make it easier and cleaner to run, manage and amend
tests. Plus, you'll be able to track all tests via git by checking the git
history of the tests file.